+++ pkg/osquery/table/platform_tables_windows.go
@@ -4,7 +4,7 @@ package table
 
 import (
 	"github.com/kolide/launcher/pkg/osquery/tables/secedit"
-	"github.com/kolide/launcher/pkg/osquery/tables/wlan"
+	"github.com/kolide/launcher/pkg/osquery/tables/wifi_networks"
 	"github.com/kolide/launcher/pkg/osquery/tables/wmitable"
 
 	"github.com/go-kit/kit/log"
@@ -17,6 +17,6 @@ func platformTables(client *osquery.ExtensionManagerClient, logger log.Logger, c
 		ProgramIcons(),
 		secedit.TablePlugin(client, logger),
 		wmitable.TablePlugin(client, logger),
-		wlan.TablePlugin(client, logger),
+		wifi_networks.TablePlugin(client, logger),
 	}
 }
