+++ pkg/osquery/tables/wifi_networks/testdata/multiple_results.txt
