+++ pkg/osquery/tables/wifi_networks/testdata/resultsps.txt
