+++ pkg/osquery/tables/wifi_networks/wifi_networks.go
@@ -1,4 +1,4 @@
-package wlan
+package wifi_networks
 
 import (
 	"bufio"
